Route to translation. Breaks window history.
